requirements.txt
streamlit==1.31.1
tensorflow>=2.8.0
keras==2.15.0
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
scikit-learn==1.3.2
pillow==10.1.0
kaggle==1.5.16
opencv-python==4.8.1.78
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5
streamlit-image-comparison==0.0.4 
